folder restructuring - scss
