+++ src/components/global/Cart/Cart.js
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import { AppContext } from "./App";
+import { AppContext } from "../../App";
 
 import CartItem from "./CartItem";
