+++ src/components/global/Navbar/Logo.js
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { ReactComponent as LogoSVG } from "../../../assets/shared/desktop/logo.svg";
-import { AppContext } from "./App";
+import { AppContext } from "../../App";
 
 export default function Logo() {
   const { handleRedirect } = useContext(AppContext);
