+++ src/components/global/Navbar/Navbar.js
@@ -1,10 +1,10 @@
 import React, { useContext } from "react";
 import { Link } from "react-router-dom";
-import { AppContext } from "./App";
+import { AppContext } from "../../App";
 
 import Logo from "./Logo";
 import MenuBtn from "./MenuBtn";
-import CategoryNav from "./CategoryNav";
+import CategoryNav from "../CategoryNav/CategoryNav";
 import { ReactComponent as ShoppingCart } from "../../../assets/shared/desktop/icon-cart.svg";
 
 export default function Navbar({ handleShowCart }) {
