+++ src/components/views/Product/ProductRecs.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ProductEachRec from "./ProductEachRec";
+import ProductRecsEach from "./ProductRecsEach";
 
 export default function ProductRecs({ product }) {
   return (
@@ -7,7 +7,7 @@ export default function ProductRecs({ product }) {
       <h3>You may also like</h3>
       <div className="product-recs-container">
         {product.others.map((rec, index) => {
-          return <ProductEachRec key={index} rec={rec} />;
+          return <ProductRecsEach key={index} rec={rec} />;
         })}
       </div>
     </section>
