+++ src/styles/css/App.css
@@ -200,7 +200,7 @@ section {
   padding: 75rem 0;
 }
 
-.checkout-success, #cart, .drop-menu {
+.checkout-success, .drop-menu, #cart {
   position: fixed;
   z-index: 99;
   top: 0;
@@ -214,78 +214,43 @@ section {
     padding: 30rem 0;
   }
 }
-#nav-bar {
-  position: fixed;
-  top: 0;
-  z-index: 99;
-  width: 100vw;
-}
-#nav-bar .nav-container {
-  padding: 0 10vw;
-}
-
-.nav-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: row;
-  background: #101010;
-  min-height: 10vh;
-  border-bottom: 1rem solid rgba(250, 250, 250, 0.5);
-}
-
-.nav-logo {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: row;
-  column-gap: 30rem;
-}
-
-.cart-btn,
-.logo {
-  cursor: pointer;
-}
-
-.drop-menu {
-  top: 10vh;
-}
-.drop-menu .category-nav {
-  padding: 50rem 10vw;
-  background: white;
-  border-radius: 0 0 10rem 10rem;
-}
-.drop-menu .categories-container {
-  row-gap: 20rem;
-}
-.drop-menu .category-nav-btns {
-  height: 200rem;
-}
-
-@media only screen and (max-width: 375px) {
-  .drop-menu .category-nav {
-    padding: 30rem 10vw;
-  }
-  .drop-menu .category-nav-btns {
-    height: 130rem;
-  }
-}
-.menu-btn {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: column;
-  width: 20rem;
-  height: 15rem;
-  cursor: pointer;
-}
-
-.white-line {
-  height: 3rem;
+.about {
   width: 100%;
-  background: white;
 }
 
+.about-grid-container {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(300rem, 1fr));
+  justify-items: center;
+  align-items: center;
+  width: 100%;
+  column-gap: 30rem;
+  row-gap: 50rem;
+  width: 100%;
+}
+
+.about-text-container {
+  width: 70%;
+}
+.about-text-container h3 span {
+  color: #d87d4a;
+}
+
+.about-image {
+  border-radius: 10rem;
+  width: 100%;
+}
+
+@media only screen and (max-width: 1250px) {
+  .about-text-container {
+    order: 2;
+    text-align: center;
+  }
+
+  .about-grid-container {
+    grid-template-columns: 1fr;
+  }
+}
 .cart-summary {
   position: absolute;
   top: 15vh;
@@ -388,6 +353,28 @@ section {
     width: 90rem;
   }
 }
+.counter {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: row;
+  width: 120rem;
+  background: #f1f1f1;
+  margin-right: 20rem;
+}
+.counter p {
+  margin: 0;
+  font-weight: bold;
+}
+.counter .counter-btn {
+  padding: 18rem 14rem;
+  color: #727272;
+  font-weight: bold;
+}
+.counter .counter-btn:hover {
+  color: #d87d4a;
+}
+
 .category-nav {
   position: relative;
   display: flex;
@@ -467,43 +454,6 @@ section {
     height: 250rem;
   }
 }
-.about {
-  width: 100%;
-}
-
-.about-grid-container {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300rem, 1fr));
-  justify-items: center;
-  align-items: center;
-  width: 100%;
-  column-gap: 30rem;
-  row-gap: 50rem;
-  width: 100%;
-}
-
-.about-text-container {
-  width: 70%;
-}
-.about-text-container h3 span {
-  color: #d87d4a;
-}
-
-.about-image {
-  border-radius: 10rem;
-  width: 100%;
-}
-
-@media only screen and (max-width: 1250px) {
-  .about-text-container {
-    order: 2;
-    text-align: center;
-  }
-
-  .about-grid-container {
-    grid-template-columns: 1fr;
-  }
-}
 footer {
   background: #101010;
   padding: 30rem 0;
@@ -581,6 +531,79 @@ footer p {
     padding: 0;
   }
 }
+#nav-bar {
+  position: fixed;
+  top: 0;
+  z-index: 99;
+  width: 100vw;
+}
+#nav-bar .nav-container {
+  padding: 0 10vw;
+}
+
+.nav-container {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: row;
+  background: #101010;
+  height: 10vh;
+  min-height: 60rem;
+  border-bottom: 1rem solid rgba(250, 250, 250, 0.5);
+}
+
+.nav-logo {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: row;
+  column-gap: 30rem;
+}
+
+.cart-btn,
+.logo {
+  cursor: pointer;
+}
+
+.drop-menu {
+  top: 10vh;
+}
+.drop-menu .category-nav {
+  padding: 50rem 10vw;
+  background: white;
+  border-radius: 0 0 10rem 10rem;
+}
+.drop-menu .categories-container {
+  row-gap: 20rem;
+}
+.drop-menu .category-nav-btns {
+  height: 200rem;
+}
+
+@media only screen and (max-width: 375px) {
+  .drop-menu .category-nav {
+    padding: 30rem 10vw;
+  }
+  .drop-menu .category-nav-btns {
+    height: 130rem;
+  }
+}
+.menu-btn {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-direction: column;
+  width: 20rem;
+  height: 15rem;
+  cursor: pointer;
+}
+
+.white-line {
+  height: 3rem;
+  width: 100%;
+  background: white;
+}
+
 header {
   position: relative;
 }
@@ -806,13 +829,6 @@ header {
   padding-top: 0;
 }
 
-.purchase-btns {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  flex-direction: row;
-}
-
 .product-features {
   display: grid;
   grid-template-columns: repeat(auto-fit, minmax(325rem, 1fr));
@@ -887,6 +903,13 @@ header {
     margin-bottom: 30rem;
   }
 }
+.purchase-btns {
+  display: flex;
+  justify-content: flex-start;
+  align-items: center;
+  flex-direction: row;
+}
+
 .product-recs h3 {
   text-align: center;
 }
@@ -913,28 +936,6 @@ header {
   border-radius: 10rem;
   width: 100%;
   margin-bottom: 40rem;
-}
-
-.counter {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-direction: row;
-  width: 120rem;
-  background: #f1f1f1;
-  margin-right: 20rem;
-}
-.counter p {
-  margin: 0;
-  font-weight: bold;
-}
-.counter .counter-btn {
-  padding: 18rem 14rem;
-  color: #727272;
-  font-weight: bold;
-}
-.counter .counter-btn:hover {
-  color: #d87d4a;
 }
 
 #checkout-page {
@@ -1042,24 +1043,6 @@ header {
     grid-template-columns: repeat(auto-fit, minmax(200rem, 1fr));
   }
 }
-.checkout-summary {
-  height: fit-content;
-}
-.checkout-summary h6 {
-  color: #727272;
-}
-.checkout-summary .orange-btn {
-  width: 100%;
-}
-
-.cart-item-container {
-  margin: 30rem 0;
-}
-
-.grand-total {
-  color: #d87d4a;
-}
-
 .checkout-success {
   left: 0;
   display: flex;
@@ -1139,4 +1122,21 @@ header {
   .success-purchase-summary {
     display: block;
   }
+}
+.checkout-summary {
+  height: fit-content;
+}
+.checkout-summary h6 {
+  color: #727272;
+}
+.checkout-summary .orange-btn {
+  width: 100%;
+}
+
+.cart-item-container {
+  margin: 30rem 0;
+}
+
+.grand-total {
+  color: #d87d4a;
 }
