+++ src/styles/scss/components/global/About/_About.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .about {
   width: 100%;
