+++ src/styles/scss/components/global/Cart/_Cart.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 #cart {
   @extend %fade-block;
