+++ src/styles/scss/components/global/Cart/_Counter.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .counter {
   @include flex-container(space-between, center);
