+++ src/styles/scss/components/global/CategoryNav/_CategoryNav.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .category-nav {
   position: relative;
