+++ src/styles/scss/components/global/CategoryNav/_CategoryNavBtn.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .category-nav-btns {
   width: 100%;
