+++ src/styles/scss/components/global/Footer/_Footer.scss
@@ -1,4 +1,4 @@
-@use "../styles/" as *;
+@use "../../../styles" as *;
 
 footer {
   background: $dark-gray;
