+++ src/styles/scss/components/global/Navbar/_MenuBtn.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .menu-btn {
   @include flex-container(space-between, center, column);
