+++ src/styles/scss/components/views/Category/_Category.scss
@@ -1,4 +1,4 @@
-@use "../../styles" as *;
+@use "../../../styles" as *;
 
 #category-page {
   padding-top: 10vh;
