+++ src/styles/scss/components/views/Category/_CategoryProduct.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .category-product {
   width: 100%;
