+++ src/styles/scss/components/views/Checkout/_Checkout.scss
@@ -1,4 +1,4 @@
-@use "../../styles" as *;
+@use "../../../styles" as *;
 
 #checkout-page {
   padding: 5vh 0;
