+++ src/styles/scss/components/views/Checkout/_CheckoutForm.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .form-section-grid {
   @include even-grid-container(275rem);
