+++ src/styles/scss/components/views/Checkout/_CheckoutSuccess.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .checkout-success {
   @extend %fade-block;
