+++ src/styles/scss/components/views/Checkout/_CheckoutSummary.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .checkout-summary {
   height: fit-content;
