+++ src/styles/scss/components/views/Home/_HomePromoItems.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 $promo-item-height: 380rem;
