+++ src/styles/scss/components/views/Product/_Product.scss
@@ -1,4 +1,4 @@
-@use "../../styles" as *;
+@use "../../../styles" as *;
 
 #product-page {
   padding-top: 15vh;
