+++ src/styles/scss/components/views/Product/_ProductGallery.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .product-gallery {
   display: grid;
