+++ src/styles/scss/components/views/Product/_ProductPurchase.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .purchase-btns {
   @include flex-container(flex-start, center);
