+++ src/styles/scss/components/views/Product/_ProductRecs.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .product-recs {
   h3 {
