+++ src/styles/scss/components/views/Product/_ProductRecsEach.scss
@@ -1,4 +1,4 @@
-@use "../styles" as *;
+@use "../../../styles" as *;
 
 .product-each-rec {
   @include flex-container(center, center, column);
